Fix the memset/memclr pointer alignment
The current code enters 8-word copy block with a 2/4 bytes
aligned pointer if the start address is 3/4-aligned,
but STM instructions require word-aligned addresses and this
has the effect of not setting/clearing the bytes that cross
into the next word at the end, if the length requires it.

The issue is that 'lsls r12, r0, #31' uses the N flag to detect
odd addresses, and the C flag to detect non-divisible-by-4 addresses.
But if the address is advanced by 1 in-between, we have to test
the address again for the case where the divisibility by 4 has been
affected by the address increment, clearing or setting the C flag.