Add optional backend for `critical-section`
This is similar to the single-threaded bare metal implementations in
`cortex-m` and `riscv`.

In theory there could be a race condition in `acquire` if an interrupt
occurs between the `read` and `write`, and the interrupt disables `IME`.
But it's probably not sensible to have an interrupt disable interrupts,
so I don't think it's necessarily to complicate things with inline
assembly using `swp`.